pyproject.toml
[project]
name = "a2a-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a>=0.44",
    "a2a-sdk>=0.3.22",
    "crewai>=1.6.1",
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
